fixed casting error in binary expr ternary operator

+++ Curt/ExprNodes.cs
@@ -66,7 +66,7 @@ namespace nodes
             {
                 switch (type)
                 {
-                    case PLUS: return (Func<object, object, object>)((p1, p2) => p1.GetType() == typeof(string) ? Concat((string)p1, (string)p2) : (int)p1 + (int)p2);
+                    case PLUS: return (Func<object, object, object>)((p1, p2) => p1.GetType() == typeof(string) ? Concat((string)p1, (string)p2) : (float)p1 + (float)p2);
                     case MINUS: return (Func<float, float, float>)((p1, p2) => p1 - p2);
                     case SLASH: return (Func<float, float, float>)((p1, p2) => p1 / p2);
                     case STAR: return (Func<float, float, float>)((p1, p2) => p1 * p2);
@@ -75,8 +75,8 @@ namespace nodes
                     case GREATER_EQUAL: return (Func<float, float, bool>)((p1, p2) => p1 >= p2);
                     case LESS: return (Func<float, float, bool>)((p1, p2) => p1 < p2);
                     case LESS_EQUAL: return (Func<float, float, bool>)((p1, p2) => p1 <= p2);
-                    case EQUAL_EQUAL: return (Func<object, object, bool>)((p1, p2) => p1.GetType() == typeof(string) ? strEquals((string)p1, (string)p2) : (int)p1 == (int)p2);
-                    case BANG_EQUAL: return (Func<object, object, bool>)((p1, p2) => p1.GetType() == typeof(string) ? strEquals((string)p1, (string)p2) : (int)p1 != (int)p2);
+                    case EQUAL_EQUAL: return (Func<object, object, bool>)((p1, p2) => p1.GetType() == typeof(string) ? strEquals((string)p1, (string)p2) : (float)p1 == (float)p2);
+                    case BANG_EQUAL: return (Func<object, object, bool>)((p1, p2) => p1.GetType() == typeof(string) ? strEquals((string)p1, (string)p2) : (float)p1 != (float)p2);
                     case AND: return (Func<bool, bool, bool>)((p1, p2) => p1 && p2);
                     case OR: return (Func<bool, bool, bool>)((p1, p2) => p1 || p2);
                     default: throw new InvalidOperationException("Unexpected binary operator");
